+++ test/nn/conv/test_gcn_conv.py
@@ -99,7 +99,12 @@ def test_static_gcn_conv():
     assert out.size() == (3, 4, 32)
 
 
-def test_gcn_conv_norm():
+def test_gcn_conv_error():
+    with pytest.raises(ValueError, match="does not support adding self-loops"):
+        GCNConv(16, 32, normalize=False, add_self_loops=True)
+
+
+def test_gcn_conv_flow():
     x = torch.randn(4, 16)
     edge_index = torch.tensor([[0, 0, 0], [1, 2, 3]])
 
@@ -108,16 +113,6 @@ def test_gcn_conv_norm():
     conv.flow = "target_to_source"
     out2 = conv(x, edge_index.flip(0))
     assert torch.allclose(out1, out2, atol=1e-6)
-
-    x = torch.tensor([[0], [2], [1]], dtype=torch.float)
-    edge_index = torch.tensor([[0, 1, 1, 2], [1, 0, 2, 1]])
-    conv = GCNConv(1, 1, add_self_loops=True, bias=False, normalize=False,
-                   aggr='max')
-    conv.lin.bias = torch.nn.Parameter(torch.tensor(0., requires_grad=True))
-    conv.lin.weight = torch.nn.Parameter(
-        torch.tensor([[1.]], requires_grad=True))
-    out3 = conv(x, edge_index)
-    assert torch.allclose(out3, torch.tensor([[2.], [2.], [2.]]), atol=1e-6)
 
 
 @pytest.mark.parametrize('requires_grad', [False, True])